+++ includes/class-tmsm-admin-cleanup.php
@@ -192,6 +192,7 @@ class Tmsm_Admin_Cleanup {
 
 		// Polylang
 		$this->loader->add_filter( 'display_post_states', $plugin_admin, 'polylang_display_post_states_language', 10, 2 );
+		$this->loader->add_action( 'parse_query', $plugin_admin, 'polylang_elementor_library_conditions_parse_query', 1, 1 );
 
 		// WP Rocket
 		$this->loader->add_filter( 'get_rocket_option_wl_plugin_name', $plugin_admin, 'wprocket_name', 10 );